refactor: decompose province_lookup for wave 5 slice D (#4125)
Split index internals and WorldStateProvinceLookup extension into
dedicated libraries; clear the repo.colonizethis_world_lib_file_size
grandfather allowlist. Update dual-region canonical path exclusions.

Refs #4125
